能够与数据库连接&通过127.0.0.1:8080/api/v1/student/book_course可以插入数据

+++ app/models/common.go
@@ -8,7 +8,7 @@ import (
 
 // 自增ID主键
 type ID struct {
-	ID uint `json:"id" gorm:"primaryKey"`
+	ID int64 `json:"id" gorm:"primaryKey"`
 }
 
 // 软删除
