+++ app/models/user_course.go
@@ -2,6 +2,6 @@ package models
 
 type UserCourse struct {
 	ID
-	UserID   uint `json:"user_id" gorm:"not null;comment:用户ID"`
-	CourseID uint `json:"course_id" gorm:"not null;comment:课程ID"`
+	UserID   int64 `json:"user_id" gorm:"not null;comment:用户ID"`
+	CourseID int64 `json:"course_id" gorm:"not null;comment:课程ID"`
 }
